Added tunedGetSystemProfilesBaseDir()
The implementation is not optimal, but it follows the
tunedGetProfilesBaseDir() implementation.

Also changed tunedGetProfilesBaseDir() to return path without the
terminating slash. I am pretty sure it wasn't originally intended,
because it leads to the following crazy usage in the scripts:

`tunedGetProfilesBaseDir`custom-profile instead of the much cleaner:
`tunedGetProfilesBaseDir`/custom-profile. If the latter is used
without this change, there will be two slashes in the logs.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>